changed sr to int64_t

+++ include/openmc/random_ray/flat_source_domain.h
@@ -100,9 +100,9 @@ protected:
     const vector<int32_t>& instances);
   void apply_external_source_to_cell_and_children(int32_t i_cell,
     Discrete* discrete, double strength_factor, int32_t target_material_id);
-  virtual void set_flux_to_flux_plus_source(int sr, double volume, int g);
-  void set_flux_to_source(int sr, int g);
-  virtual void set_flux_to_old_flux(int sr, int g);
+  virtual void set_flux_to_flux_plus_source(int64_t sr, double volume, int g);
+  void set_flux_to_source(int64_t sr, int g);
+  virtual void set_flux_to_old_flux(int64_t sr, int g);
 
   //----------------------------------------------------------------------------
   // Private data members
